loading data by useEffect and refactoring DB to virtual server
