+++ FAKE_DATA/DB.json
@@ -1,0 +1,197 @@
+{"catalogItems":[
+  {
+   "title": "Medicines and prophylactics",
+   "img": "https://root.tblcdn.com/icons/category/c700d3d9-a353-438a-aebb-6906bf42e10a/img_0.png?v=AAAAAAnmWC0",
+   "content": [
+      {
+       "title": "Nervous System",
+      "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Anxiolytics","subLink": "#" },
+          {"subTitle": "Antidepressants","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Painkillers",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Analgesics","subLink": "#" },
+          {"subTitle": "Anesthesia, anesthetics in dentistry","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Nervous System",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Anxiolytics","subLink": "#" },
+          {"subTitle": "Antidepressants","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Painkillers",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Analgesics","subLink": "#" },
+          {"subTitle": "Anesthesia, anesthetics in dentistry","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Nervous System",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Anxiolytics","subLink": "#" },
+          {"subTitle": "Antidepressants","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Painkillers",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Analgesics","subLink": "#" },
+          {"subTitle": "Anesthesia, anesthetics in dentistry","subLink": "#" }
+        ]
+      }
+    ]
+  },
+  {
+   "title": "Vitamins and minerals",
+   "img": "https://root.tblcdn.com/icons/category/ba1fb752-c42f-4580-a3d7-37b5ab977e74/img_0.png?v=AAAAAAnmWHo",
+   "content": [
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Vitamins",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Biotin","subLink": "#" },
+          {"subTitle": "Vitamin D","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Minerals",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Iron (Fe)","subLink": "#" },
+          {"subTitle": "Йод (I)","subLink": "#" }
+        ]
+      }
+    ]
+  },
+  {
+   "title": "Beauty and care",
+   "img": "https://root.tblcdn.com/icons/category/c8b01654-b691-4bf2-8d2a-0da811ccd1e2/img_0.png?v=AAAAAAox8GI",
+   "content": [
+      {
+       "title": "Tanning agent and sun protection",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Protection for children","subLink": "#" },
+          {"subTitle": "After tanning","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Facial",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Антивіковий погляд","subLink": "#" },
+          {"subTitle": "Бальзами для губ","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Tanning agent and sun protection",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Protection for children","subLink": "#" },
+          {"subTitle": "After tanning","subLink": "#" }
+        ]
+      },
+      {
+       "title": "Facial",
+       "titleLink": "#",
+       "subCategories": [
+          {"subTitle": "Антивіковий погляд","subLink": "#" },
+          {"subTitle": "Бальзами для губ","subLink": "#" }
+        ]
+      }
+    ]
+  }
+]
+}
