+++ FAKE_DATA/catalogItems.ts
