+++ FAKE_DATA/cities.ts
