+++ FAKE_DATA/richCategories.ts
