+++ package.json
@@ -6,10 +6,12 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "json-server": "json-server --watch FAKE_DATA/DB.json --port 3005"
   },
   "dependencies": {
     "autoprefixer": "10.4.16",
+    "axios": "^1.6.2",
     "next": "13.5.2",
     "postcss": "8.4.30",
     "react": "18.2.0",
@@ -29,6 +31,7 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "json-server": "^0.17.4",
     "prettier": "^3.0.3",
     "typescript": "^5.2.2"
   }
