+++ src/components/catalog/CatalogContentItem.tsx
@@ -16,11 +16,13 @@ const CatalogContentItem: FC<ICatalogContent> = ({
         <Link href={titleLink}>{title}</Link>
       </h3>
       <ul>
-        {subCategories.map((categ) => (
-          <li key={v4()}>
-            <Link href={categ.subLink}>{categ.subTitle}</Link>
-          </li>
-        ))}
+        {subCategories.map((categ) => {
+          return (
+            <li key={v4()}>
+              <Link href={categ.subLink}>{categ.subTitle}</Link>
+            </li>
+          );
+        })}
       </ul>
     </div>
   );
